Combine slm wrapper and abstraction.
